requirements.txt
appnope==0.1.2
atlassian-python-api==3.14.1
backcall==0.2.0
certifi==2021.5.30
charset-normalizer==2.0.6
decorator==5.1.0
Deprecated==1.2.13
idna==3.2
ipython==7.28.0
jedi==0.18.0
matplotlib-inline==0.1.3
oauthlib==3.1.1
parso==0.8.2
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.20
ptyprocess==0.7.0
Pygments==2.10.0
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
traitlets==5.1.0
urllib3==1.26.7
wcwidth==0.2.5
wrapt==1.12.1
